move some logging to dedicated subsystem loggers
