+++ hibernate-core/src/main/java/org/hibernate/stat/internal/StatisticsInitiator.java
@@ -7,8 +7,6 @@ package org.hibernate.stat.internal;
 import org.hibernate.HibernateException;
 import org.hibernate.boot.registry.classloading.spi.ClassLoaderService;
 import org.hibernate.engine.config.spi.ConfigurationService;
-import org.hibernate.internal.CoreLogging;
-import org.hibernate.internal.CoreMessageLogger;
 import org.hibernate.service.spi.ServiceRegistryImplementor;
 import org.hibernate.service.spi.SessionFactoryServiceInitiator;
 import org.hibernate.service.spi.SessionFactoryServiceInitiatorContext;
@@ -20,13 +18,12 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 
 
 import static org.hibernate.cfg.StatisticsSettings.STATS_BUILDER;
+import static org.hibernate.internal.log.StatisticsLogger.STATISTICS_LOGGER;
 
 /**
  * @author Steve Ebersole
  */
 public class StatisticsInitiator implements SessionFactoryServiceInitiator<StatisticsImplementor> {
-
-	private static final CoreMessageLogger LOG = CoreLogging.messageLogger( StatisticsInitiator.class );
 
 	public static final StatisticsInitiator INSTANCE = new StatisticsInitiator();
 
@@ -43,7 +40,7 @@ public class StatisticsInitiator implements SessionFactoryServiceInitiator<Stati
 		final var statisticsFactory = statisticsFactory( configValue, context.getServiceRegistry() );
 		final var statistics = statisticsFactory.buildStatistics( context.getSessionFactory() );
 		final boolean enabled = context.getSessionFactoryOptions().isStatisticsEnabled();
-		LOG.statisticsInitialized();
+		STATISTICS_LOGGER.statisticsInitialized();
 		statistics.setStatisticsEnabled( enabled );
 		return statistics;
 	}
